Initial Support for Spring-Security-OAuth2.0
This version depends on building a custom branch of cas that includes
spring security oauth support. This is because of custom endpoints
defined in there. Converted the demo server to use the war overlay
method of including cas server support. At this point we still need to
get the client working.