Fixed the session expiration issue
The terminate web session listener was killing our session store that held our authorizationRequest for spring-security-oauth. By default it was killing our session in 2 seconds. I commented it out to allow the session to use its default timeout (5 minutes according to web.xml).
After 5 minutes we do get the error page that the authorizationRequest is uninitialized, so it is a bug, just not as huge of a one.
Added 2 entries to the deployer config, but we might as well make it 1 in the future (or 3 explicit endpoints, depending on how secure we need it.)
The OAuth config now is slightly better documented, though there's plenty of room for improvement. Made the c and p namespace attributes more uniform.
Added cas client support for authentication, except for http basic authentication in the resource owner password grant type so the client can provide the client_id and the client_secret in a secure way.