Working changes to utilize enhanced pac4j support
Removed redundancies in the configs and made them match as much as possible to
the originals before they were modified. The spring web project had to be
excluded because of a versioning conflict with another artifact that maven
didn't/couldn't resolve. The oauthConfiguration XML file should be made generic
and placed in the cas unused spring configuration directory in
cas-server-webapp.